Code/RDBoost/Exceptions.h moved to Code/RDGeneral
