+++ Code/RDGeneral/Exceptions.h
